+++ tests/test_triples_factory.py
@@ -17,7 +17,6 @@ import torch
 from pykeen.datasets import Hetionet, Nations, SingleTabbedDataset
 from pykeen.datasets.nations import NATIONS_TRAIN_PATH
 from pykeen.triples import CoreTriplesFactory, TriplesFactory, TriplesNumericLiteralsFactory, generation
-from pykeen.triples.instances import BatchedSLCWAInstances
 from pykeen.triples.splitting import splitter_resolver
 from pykeen.triples.triples_factory import (
     INVERSE_SUFFIX,
@@ -81,17 +80,6 @@ class TestTriplesFactory(unittest.TestCase):
     def setUp(self) -> None:
         """Instantiate test instance."""
         self.factory = Nations().training
-
-    def test_correct_inverse_creation(self):
-        """Test if the triples and the corresponding inverses are created."""
-        t = [
-            ["e1", "a.", "e5"],
-            ["e1", "a", "e2"],
-        ]
-        t = np.array(t, dtype=str)
-        factory = TriplesFactory.from_labeled_triples(triples=t, create_inverse_triples=True)
-        instances = BatchedSLCWAInstances.from_triples_factory(factory)
-        assert len(instances) == 4
 
     def test_automatic_incomplete_inverse_detection(self):
         """Test detecting that the triples contain inverses, warns about them, and filters them out."""
